decoder: Prevent signed/unsigned comparison in rlefont
Adding a signed and unsigned 16 bit value in C promotes the signed value
to unsigned, losing the sign bit in the process. Comparing then with an
signed value can cause potential bugs and at least throws a warning when
compiling when `sign-compare` is enabled.

This patch casts the int16_t first to int32_t to prevent the implicit
cast to uint16_t.