decoder: Change char width to signed
This patch changes the type of the char width in append_char to int16_t.
As mf_character_width returns uint8_t, this is safe. By changing the
type to int16_t, an implicit conversion to uint16_t in the addition
later is prevented and in turn a comparision between signed and unsigned
types is prevented.